mosaic_bot/requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.0
Django==3.1.2
django-jinja==2.7.0
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.10
yarl==1.6.0
